backend/pyproject.toml
[project]
name = "discipleship-backend"
version = "0.1.0"
description = "Discipleship API — mentor matching"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "sqlalchemy>=2.0",
  "psycopg[binary]>=3.1",
  "pgvector>=0.2.5",
  "alembic>=1.13",
  "pydantic>=2.6",
  "pydantic-settings>=2.2",
  "email-validator>=2.1",
  "python-jose[cryptography]>=3.3",
  "passlib[bcrypt]>=1.7",
  "sentence-transformers>=2.6",
  "numpy>=1.26",
  "python-multipart>=0.0.9",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
